pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chowkidar"
version = "0.1.0"
description = "Local-first LLM model deprecation watchdog — scans your configs, alerts on sunsets, and instructs your IDE to update."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Bhavishya" }]
keywords = ["llm", "deprecation", "model", "watchdog", "mcp", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "plyer>=2.1.0",
    "filelock>=3.12.0",
    "pyyaml>=6.0",
    "schedule>=1.2.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
slm = ["ollama>=0.4.0"]
dev = [
    "pytest>=7.0",
    "respx>=0.20.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
chowkidar = "chowkidar.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/chowkidar"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
